All aspects of the tooltip are now working with using the store to store the time locations of the tooltips.
