+++ assets/src/scripts/components/hydrograph/layout.js
@@ -5,7 +5,7 @@ const { createSelector } = require('reselect');
 const ASPECT_RATIO = 1 / 2;
 const ASPECT_RATIO_PERCENT = `${100 * ASPECT_RATIO}%`;
 const MARGIN = {
-    top: 35,
+    top: 40,
     right: 75,
     bottom: 50,
     left: 50
