+++ assets/src/styles/components/_hydrograph.scss
@@ -39,7 +39,6 @@ $estimated: black;
         .tooltip-text-group {
             .current-tooltip-text {
                 font-weight: bold;
-                fill: red;
             }
             .approved {
                 fill: $approved;
